Added test controller
Added a test controller that can be called from the launch file. The controller moves all of the joints simultaneously, but has legacy functions for single joint movement. To launch the test controller specify controller:=test_no_comp for the controller without friction compensation, and test_comp for the controller with friction compensation.